got error and reseting the commit
