+++ teacher.rb
@@ -1,23 +1,18 @@
-class Teacher
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
-  attr_accessor :name
+require_relative 'employee'
 
-  def initialize(options={})
+class Teacher < Employee
+  attr_reader :salary, :phase, :performance_rating, :target_raise
+  
+  TARGET_RAISE = 1000
+
+  def initialize(options={}) 
+    super
+    @target_raise = TARGET_RAISE
     @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 1000
   end
-
-  def offer_high_five
-    "High five!"
+    def receive_raise(raise)
+    @salary += raise
   end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
-
   def teach_stuff
     response = ""
     response += "Listen, class, this is how everything works, fo SHO! "
@@ -26,19 +21,10 @@ class Teacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
   def set_performance_rating(rating)
     response = ""
     if rating > 90
-      receive_raise(@target_raise)
+      receive_raise(TARGET_RAISE)
       response = "Yay, I'm a great employee!"
     else
       response += "Oh, well -- thanks to this actionable, specific, and kind "
